requirements.txt
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
nltk==3.8.1  # Optional: for better text processing